# book-management/src/main/resources/swagger/book-management.yaml
openapi: "3.0.4"
info:
  title: "Book Management API"
  version: "1.0.0"
  description: "API para gerenciar livros"
servers:
  - url: "http://localhost:8080"
    description: "Ambiente local"
...
paths:
  /api/v1/books:
    get:
      tags:
        - Books
      summary: "List of Books"
      operationId: getAllBooks
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/NameQuery'
      responses:
        '200':
          description: "Book List"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponseData'
    post:
      tags:
        - Books
      summary: "Create a new Book"
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequestDTO'
      responses:
        '201':
          description: "Book created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponseData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/books/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags:
        - Books
      summary: "Find book by ID"
      operationId: getBookById
      responses:
        '200':
          description: "Book found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponseData'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Books
      summary: "Update book by ID"
      operationId: updateBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequestDTO'
      responses:
        '200':
          description: "Book updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponseData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Books
      summary: "Delete book"
      operationId: deleteBook
      responses:
        '204':
          description: "Book deleted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    Page:
      name: "page"
      in: "query"
      description: "Page"
      required: false
      schema:
        type: integer
        format: int32
        default: 0
    Size:
      name: "size"
      in: "query"
      description: "Size"
      required: false
      schema:
        type: integer
        format: int32
        default: 10
    IdParam:
      name: "id"
      in: "path"
      description: "Book ID"
      required: true
      schema:
        type: integer
        format: int64
    NameQuery:
      name: "name"
      in: "query"
      description: "Name"
      required: false
      schema:
        type: string
        format: string
        default: ""
...
  schemas:
    BookRequestDTO:
      type: object
      properties:
        bookCode:
          type: integer
          format: int64
        title:
          type: string
        publisher:
          type: string
        edition:
          type: string
        yearPublication:
          type: string
          example: 1999
        price:
          type: string
          format: decimal
          example: 100.00
        authorsIds:
          type: array
          items:
            type: integer
        subjectsIds:
          type: array
          items:
            type: integer
      required:
        - bookCode
        - title
        - publisher
        - edition
        - yearPublication
        - price
        - authorsIds
        - subjectsIds
...
    BookResponseData:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookResponseDTO'
    BookResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookCode:
          type: string
          example: "BK-123456"
        title:
          type: string
        publisher:
          type: string
        edition:
          type: string
        yearPublication:
          type: string
          example: "2023"
        price:
          type: string
          format: decimal
        authors:
          type: array
          items:
            type: integer
        subjects:
          type: array
          items:
            type: integer
      required:
        - bookCode
        - title
        - publisher
        - edition
        - yearPublication
        - price
        - authors
        - subjects
...
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        message:
          type: string
        path:
          type: string
      required: ["code", "message"]
  responses:
    BadRequest:
      description: "Requisição inválida"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: "Recurso não encontrado"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: "Erro interno do servidor"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'